+++ src/Concordium/Client/LegacyCommands.hs
@@ -160,6 +160,10 @@ data LegacyCmd
         {legacyBlockHash :: !(Maybe Text)}
     | GetTokenList
         {legacyBlockHash :: !(Maybe Text)}
+    | GetTokenInfo
+        { tokenId :: !Text,
+          legacyBlockHash :: !(Maybe Text)
+        }
     | GetBakersRewardPeriod
         {legacyBlockHash :: !(Maybe Text)}
     | GetBlockCertificates
@@ -228,6 +232,7 @@ legacyProgramOptions =
             <> getCryptographicParametersCommand
             <> getNextUpdateSequenceNumbersCommand
             <> getTokenListCommand
+            <> getTokenInfoCommand
             <> getScheduledReleaseAccountsCommand
             <> getCooldownAccountsCommand
             <> getPreCooldownAccountsCommand
@@ -454,6 +459,18 @@ getTokenListCommand =
                 <$> optional (strArgument (metavar "BLOCK-HASH" <> help "Hash of the block to query (default: Query the best block)"))
             )
             (progDesc "Query the gRPC server for the protocol level token list.")
+        )
+
+getTokenInfoCommand :: Mod CommandFields LegacyCmd
+getTokenInfoCommand =
+    command
+        "GetTokenInfo"
+        ( info
+            ( GetTokenInfo
+                <$> strOption (long "tokenId" <> metavar "TOKEN_ID" <> help "Token id (Symbol) of the token.")
+                <*> optional (strArgument (metavar "BLOCK-HASH" <> help "Hash of the block to query (default: Query the best block)"))
+            )
+            (progDesc "Query the gRPC server for the info of a protocol level token.")
         )
 
 getScheduledReleaseAccountsCommand :: Mod CommandFields LegacyCmd